requirements.txt
openai>=1.0.0
anthropic>=0.32.0
requests>=2.31.0
python-dotenv>=1.0.0
moviepy>=1.0.3
Pillow>=10.0.0
